+++ common-lib/esm/http.js
@@ -67,17 +67,22 @@ const http = async (method, url, body = null, query = null, headers = null) => {
 
     if (!headers) {
       headers = {
-        Accept: 'application/json',
-        'Content-Type': 'application/json'
+        Accept: 'application/json'
       }
     }
     const options = { method, headers }
     if (timeoutMs > 0) options.signal = signal
     if (token && credentials !== 'include') options.headers.Authorization = `Bearer ${token}` // include === HTTPONLY_TOKEN
     if (urlPath === '/api/auth/logout') options.headers.refresh_token = refreshToken // add refresh token for logout
-    if (body) options.body = JSON.stringify(body)
+    if (body) {
+      if (body instanceof FormData) {
+        options.body = body
+      } else {
+        headers['Content-Type'] = 'application/json'
+        options.body = JSON.stringify(body)
+      }
+    }
     options.credentials = credentials
-
     const rv0 = await fetch(urlFull + qs, options)
     // rv0.data = await rv0.json() // replaced by below to handle empty body
     const txt0 = await rv0.text()